refactor: breakdown DropZone render into hooks
